9단계 구현, 권한 확인 구현, 일정 수정 및 삭제는 관리자 권한이 있는 유저만, 예외처리 권한없으면 403error 반환

+++ src/main/java/com/sparta/jpaupgradeschedule/service/ScheduleService.java
@@ -3,10 +3,15 @@ package com.sparta.jpaupgradeschedule.service;
 import com.sparta.jpaupgradeschedule.dto.*;
 import com.sparta.jpaupgradeschedule.entity.Schedule;
 import com.sparta.jpaupgradeschedule.entity.User;
+import com.sparta.jpaupgradeschedule.entity.UserRoleEnum;
 import com.sparta.jpaupgradeschedule.entity.UserSchedule;
+import com.sparta.jpaupgradeschedule.jwt.JwtUtil;
 import com.sparta.jpaupgradeschedule.repository.ScheduleRepository;
 import com.sparta.jpaupgradeschedule.repository.UserRepository;
 import com.sparta.jpaupgradeschedule.repository.UserScheduleRepository;
+import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -15,6 +20,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
@@ -25,8 +31,10 @@ import java.util.Set;
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
-    private final UserService userService;
+    private final UserRepository userRepository;
     private final UserScheduleRepository userScheduleRepository;
+    private final UserService userService;
+    private final JwtUtil jwtUtil;
 
     @Transactional
     public ScheduleSaveResponseDto saveSchedule(ScheduleSaveRequestDto requestDto, Set<User> user) {
@@ -62,11 +70,17 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleSaveResponseDto updateSchedule(Long id, ScheduleSaveRequestDto requestDto) {
-        Schedule idSchedule = scheduleFindById(id);
-        idSchedule.update(requestDto);
+    public ScheduleSaveResponseDto updateSchedule(Long id, ScheduleSaveRequestDto requestDto, HttpServletResponse res, HttpServletRequest req) throws IOException {
+        Claims info = jwtUtil.getUserInfoFromToken(jwtUtil.substringToken(jwtUtil.getTokenFromRequest(req)));
+        if (userRepository.findByEmail(info.getSubject()).getRole().equals(UserRoleEnum.ADMIN)) {
+            Schedule idSchedule = scheduleFindById(id);
+            idSchedule.update(requestDto);
 
-        return new ScheduleSaveResponseDto(idSchedule);
+            return new ScheduleSaveResponseDto(idSchedule);
+        } else {
+            res.sendError(403, "403, role is not ADMIN");
+            return null;
+        }
     }
 
     public Page<SchedulePageResponseDto> getSchedules(int page, int size) {
@@ -80,9 +94,17 @@ public class ScheduleService {
     }
 
     @Transactional
-    public String deleteSchedule(Long id) {
-        scheduleRepository.delete(scheduleFindById(id));
-        return "삭제완료";
+    public String deleteSchedule(Long id, HttpServletResponse res, HttpServletRequest req) throws IOException {
+        Claims info = jwtUtil.getUserInfoFromToken(jwtUtil.substringToken(jwtUtil.getTokenFromRequest(req)));
+        if (userRepository.findByEmail(info.getSubject()).getRole().equals(UserRoleEnum.ADMIN)) {
+            List<UserSchedule> userScheduleList = userScheduleRepository.findAllBySchedule(scheduleFindById(id));
+            userScheduleRepository.deleteAll(userScheduleList);
+            scheduleRepository.delete(scheduleFindById(id));
+            return "삭제완료";
+        } else {
+            res.sendError(403, "role is not ADMIN");
+            return "role is not ADMIN";
+        }
     }
 
     public Schedule scheduleFindById(Long id) {
